pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sql-docs"
version = "0.1.0"
description = "Generate HTML documentation from SQL queries"
authors = [
    { name = "Emiliano Rigueiro" }
]
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "jinja2",
    "pyyaml",
    "pandas"
]

keywords = ["sql", "documentation", "data-engineering"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sqldocs*"]

[tool.setuptools.package-data]
"sqldocs" = ["templates/*.html"]